+++ src/validator/verification/workchain.rs
@@ -47,10 +47,8 @@ use metrics::Recorder;
 use catchain::utils::add_compute_relative_metric;
 use catchain::utils::add_compute_result_metric;
 
-//TODO: response to MC/WC node if this node has extra BLS information compared to sender node (do not send back if received from MC overlay)
-//TODO: sync time should be affected by last update time of the block status
-//TODO: load empty block statuses at workchain start (top of shards) to speed up recovery process after restart
-//TODO: test sync time is the same as lifetime and sync until block status is delivered
+//TODO: add cutoff to limit number of WC validators which can send block to MC during one time slot
+//TODO: send block status to MC via broadcast
 
 /*
 ===============================================================================
@@ -77,6 +75,7 @@ pub struct Workchain {
     wc_validators: Vec<ValidatorDescr>,   //WC validators
     mc_validators: Vec<ValidatorDescr>,   //MC validators
     wc_validators_adnl_ids: Vec<PublicKeyHash>, //WC validators ADNL IDs
+    mc_validators_adnl_ids: Vec<PublicKeyHash>, //MC validators ADNL IDs
     wc_pub_keys: Vec<[u8; BLS_PUBLIC_KEY_LEN]>, //WC validators pubkeys
     local_adnl_id: PublicKeyHash,         //ADNL ID for this node
     wc_local_idx: i16,                    //local index in WC validator set
@@ -291,6 +290,7 @@ impl Workchain {
             runtime: runtime.clone(),
             wc_validators_adnl_ids: wc_validators.iter().map(|desc| get_adnl_id(&desc)).collect(),
             wc_validators,
+            mc_validators_adnl_ids: mc_validators.iter().map(|desc| get_adnl_id(&desc)).collect(),
             mc_validators: mc_validators.clone(),
             wc_validator_set_hash,
             mc_validator_set_hash,
@@ -393,7 +393,36 @@ impl Workchain {
             *workchain.workchain_overlay.lock() = Some(workchain_overlay);
         }
 
+        //init workchain with top shard blocks for initial synchronization
+
+        workchain.init_top_shard_blocks();
+
         Ok(workchain)
+    }
+
+    /// Initial synchronization
+    fn init_top_shard_blocks(&self) {
+        let workchain = self.get_self();
+        let engine = self.engine.clone();
+        let runtime = workchain.runtime.clone();
+
+        runtime.spawn(async move {
+            log::debug!(target: "verificator", "Init top shard blocks for workchain {}", workchain.node_debug_id);
+            match engine.load_last_applied_mc_state().await {
+                Ok(mc_state) => {
+                    match mc_state.top_blocks(workchain.workchain_id) {
+                        Ok(top_shard_blocks) => {
+                            for block_id in top_shard_blocks {
+                                log::trace!(target: "verificator", "Init top shard block {} for workchain {}", block_id, workchain.node_debug_id);
+                                workchain.load_block_candidate(&block_id);
+                            }
+                        },
+                        Err(err) => log::error!(target: "verificator", "Can't get top shard blocks: {:?} (workchain={})", err, workchain.node_debug_id),
+                    }
+                },
+                Err(err) => log::error!(target: "verificator", "Can't load last applied MC state: {:?} (workchain={})", err, workchain.node_debug_id),
+            }
+        });
     }
 
     /*
@@ -658,43 +687,46 @@ impl Workchain {
         result.append(format!("  - block_sync_lifetime_period: {:.3}s\n", Duration::from_millis(config_params.block_sync_lifetime_period_ms as u64).as_secs_f64()));
         result.append(format!("  - blocks:\n"));
 
-        for (i, (first_appearance_time, block)) in blocks.iter().enumerate() {
-            let life_time = match first_appearance_time.elapsed() {
-                Ok(elapsed) => elapsed.as_secs_f64(),
-                Err(_) => 0.0,
-            };
-            let block = block.lock();
-            let delivered_weight = block.get_deliveries_signature().get_total_weight(&self.wc_validators);
-
-            result.append(format!("      - b{:03} ({:6.2}%): {}{}{}{}{} | delivery_latency={:.3}s, lifetime={:.3}s, hops={:2}, acks={:?}, nacks={:?}, stats={:?}, id={}\n",
-                i,
-                delivered_weight as f64 / self.wc_total_weight as f64 * 100.0,
-                if block.is_delivered(&self.wc_validators, delivery_params.wc_cutoff_weight) { " " } else { "↔" },
-                if block.is_sent_to_mc() { "↑" } else { " " },
-                if block.is_received_from_mc() { "↓" } else { " " },
-                if block.is_rejected() { "N" } else { " "},
-                if block.has_approves() { "A" } else { " " },
-                if let Some(state_change_time) = block.get_delivery_state_change_time() { 
-                    if let Some(creation_time) = block.get_creation_time() {
-                        if let Ok(latency) = state_change_time.duration_since(creation_time) {
-                            latency.as_secs_f64()
-                        } else {
-                            0.0 
-                        }
-                    } else {
-                        0.0
-                    }
-                } else { 
-                    0.0 
-                },
-                life_time,
-                block.get_merges_count(),
-                block.get_approvals_signature(),
-                block.get_rejections_signature(),
-                block.get_delivery_stats(),
-                block.get_id(),
-            ));
+        for (i, (_first_appearance_time, block)) in blocks.iter().enumerate() {
+            result.append(format!("      - b{:03} {}\n", i, self.get_readable_block_info(&block)));
         }
+    }
+
+    fn get_readable_block_info(&self, block: &BlockPtr) -> String {
+        let delivery_params = self.workchain_delivery_params.lock().clone();
+        let config_params = &delivery_params.config_params;        
+        let block = block.lock();
+        let life_time = block.get_first_appearance_time().elapsed().unwrap_or(Duration::from_secs(0)).as_secs_f64();
+        let delivered_weight = block.get_deliveries_signature().get_total_weight(&self.wc_validators);
+
+        format!("({:6.2}%): {}{}{}{}{}{} | delivery_latency={:.3}s, lifetime={:.3}s, hops={:2}, acks={:?}, nacks={:?}, stats={:?}, id={}\n",
+            delivered_weight as f64 / self.wc_total_weight as f64 * 100.0,
+            if block.is_delivered(&self.wc_validators, delivery_params.wc_cutoff_weight) { " " } else { "↔" },
+            if block.is_sent_to_mc() { "↑" } else { " " },
+            if block.is_delivered_and_received_from_mc() { "↓" } else { " " },
+            if block.is_rejected() { "N" } else { " "},
+            if block.has_approves() { "A" } else { " " },
+            if block.is_synchronizing(Duration::from_millis(config_params.block_sync_lifetime_period_ms as u64)) { "S" } else { " " },
+            if let Some(state_change_time) = block.get_delivery_state_change_time() { 
+                if let Some(creation_time) = block.get_creation_time() {
+                    if let Ok(latency) = state_change_time.duration_since(creation_time) {
+                        latency.as_secs_f64()
+                    } else {
+                        0.0 
+                    }
+                } else {
+                    0.0
+                }
+            } else { 
+                0.0 
+            },
+            life_time,
+            block.get_merges_count(),
+            block.get_approvals_signature(),
+            block.get_rejections_signature(),
+            block.get_delivery_stats(),
+            block.get_id(),
+        )
     }
 
     /*
@@ -781,7 +813,12 @@ impl Workchain {
                 None => {
                     trace!(target: "verificator", "Creating new block {:?} for workchain {}", block_id, self.node_debug_id);
 
-                    let new_block = Block::create(block_id, &*self.blocks_instance_counter, self.wc_validators.len());
+                    let new_block = Block::create(
+                        block_id,
+                        &*self.blocks_instance_counter,
+                        self.wc_validators.len(),
+                        self.mc_validators.len()
+                    );
 
                     blocks.insert(Self::get_candidate_id(block_id), new_block.clone());
 
@@ -831,7 +868,12 @@ impl Workchain {
     }
 
     /// Block update function
-    fn synchronize_block(workchain_weak: Weak<Workchain>, block_weak: Weak<SpinMutex<Block>>, far_neighbours_sync_time: Option<SystemTime>, mut manual_candidate_loading_time: Option<SystemTime>) {
+    fn synchronize_block(
+        workchain_weak: Weak<Workchain>,
+        block_weak: Weak<SpinMutex<Block>>,
+        far_neighbours_sync_time: Option<SystemTime>,
+        mut manual_candidate_loading_time: Option<SystemTime>,
+    ) {
         let workchain = {
             if let Some(workchain) = workchain_weak.upgrade() {
                 workchain
@@ -849,7 +891,7 @@ impl Workchain {
         let delivery_params = workchain.workchain_delivery_params.lock().clone();
         let config_params = &delivery_params.config_params;
 
-        let (block_id, block_end_of_life_time, block_end_of_sync_time, delivery_weight) = {
+        let (block_id, block_end_of_life_time, is_synchronizing, delivery_weight) = {
             let block = block.lock();
             let delivered_weight = block.get_deliveries_signature().get_total_weight(&workchain.wc_validators);
 
@@ -865,7 +907,7 @@ impl Workchain {
 
             (block.get_id().clone(),
              *block.get_first_appearance_time() + Duration::from_millis(config_params.block_lifetime_period_ms as u64),
-             block_start_sync_time + Duration::from_millis(config_params.block_sync_lifetime_period_ms as u64),
+             block.is_synchronizing(Duration::from_millis(config_params.block_sync_lifetime_period_ms as u64)),
              delivered_weight
             )
         };
@@ -915,7 +957,7 @@ impl Workchain {
 
         //sync with far neighbours
 
-        if block_end_of_sync_time.elapsed().is_err() {
+        if is_synchronizing {
             //trace!(target: "verificator", "Synchronize block {:?} for workchain {}", candidate_id, workchain.node_debug_id);
             trace!(target: "verificator", "Synchronize block {:?} for workchain {}: {:?}", block_id, workchain.node_debug_id, block.lock());
 
@@ -946,6 +988,12 @@ impl Workchain {
         let ready_to_send = block.lock().toggle_send_ready(false);
 
         if ready_to_send {
+            //update start sync time (because block status has been changed)
+
+            block.lock().awake_synchronization();
+
+            //send block status to neighbours
+
             workchain.send_block_status_to_forwarding_neighbours(&block, delivery_params.block_status_forwarding_neighbours_count);
         }        
 
@@ -993,7 +1041,7 @@ impl Workchain {
     fn merge_block_status(
         &self,
         block_id: &BlockIdExt,
-        source_node_idx: Option<usize>,
+        source_node_adnl_id: &PublicKeyHash,
         deliveries_signature: &MultiSignature,
         approvals_signature: &MultiSignature,
         rejections_signature: &MultiSignature,
@@ -1021,27 +1069,89 @@ impl Workchain {
 
         //update node's knowledge of the block
 
-        if let Some(source_node_idx) = source_node_idx {
-            let source_node_delivery_weight = deliveries_signature.get_total_weight(&self.wc_validators);
+        let source_node_delivery_weight = deliveries_signature.get_total_weight(&self.wc_validators);        
+
+        if let Some(source_wc_node_idx) = self.get_wc_validator_idx_by_adnl_id(source_node_adnl_id) {
             let mut block = block.lock();
 
-            block.set_node_delivery_weight(source_node_idx, source_node_delivery_weight);
-            block.set_node_status_received_time(source_node_idx, SystemTime::now());
+            block.set_wc_node_delivery_weight(source_wc_node_idx, source_node_delivery_weight);
+            block.set_wc_node_status_received_time(source_wc_node_idx, SystemTime::now());
         }
 
-        //check if block is MC originated
+        if let Some(source_mc_node_idx) = self.get_mc_validator_idx_by_adnl_id(source_node_adnl_id) {
+            let mut block = block.lock();
 
-        if !received_from_workchain {
-            block.lock().mark_as_received_from_mc();
+            block.set_mc_node_delivery_weight(source_mc_node_idx, source_node_delivery_weight);
+            block.set_mc_node_status_received_time(source_mc_node_idx, SystemTime::now());
         }
 
         //update status
 
-        let status = block.lock().merge_status(deliveries_signature, approvals_signature, rejections_signature, merges_count, created_timestamp
-            );
+        let status = block.lock().merge_status(deliveries_signature, approvals_signature, rejections_signature, merges_count, created_timestamp);
         match status {
-            Ok(status) => {
-                if !status {
+            Ok((self_status_changed, other_status_should_be_changed)) => {
+                if other_status_should_be_changed {
+                    //other node status should be updated
+
+                    if !received_from_workchain {
+                        //send block status back to MC overlay
+                        //do not answer for internal WC messages; far neighbours update approach will update WC soon
+
+                        let delivery_params = self.workchain_delivery_params.lock().clone();
+
+                        let is_delivered = block.lock().is_delivered(&self.wc_validators, delivery_params.wc_cutoff_weight);
+
+                        if is_delivered {
+                            let is_delivered_on_other_node = source_node_delivery_weight >= self.get_wc_cutoff_weight();
+
+                            if !is_delivered_on_other_node {
+                                //send back to MC only blocks which are delivered for this node and not delivered on MC node
+
+                                if let Some(source_mc_node_idx) = self.get_mc_validator_idx_by_adnl_id(source_node_adnl_id) {
+                                    //send back delivered blocks to MC only
+
+                                    let next_allowed_send_back_time = match block.lock().get_mc_node_status_sent_time(source_mc_node_idx) {
+                                        Some(last_send_back_time) => last_send_back_time + Duration::from_millis(delivery_params.config_params.mc_allowed_force_delivery_delay_ms as u64),
+                                        None => SystemTime::now(),
+                                    };
+
+                                    if next_allowed_send_back_time.elapsed().is_ok() {
+                                        //manage timeout between sends
+
+                                        if log_enabled!(log::Level::Debug)
+                                        {        
+                                            log::debug!(target: "verificator", "Block {:?} status was sent back to MC overlay node [{}]={} in workchain {}. BlockInfo: delivery={}",
+                                                block_id,
+                                                source_mc_node_idx,
+                                                source_node_adnl_id,
+                                                self.node_debug_id,
+                                                self.get_readable_block_info(&block));
+                                        }
+                
+                                        self.send_block_status_to_mc_node(&block, source_mc_node_idx);                                        
+                                    } else {
+                                        log::trace!(target: "verificator", "Block {:?} status is not sent back to MC overlay node [{}]={} in workchain {} because of timeout (next allowed send time is in {:.3}s)",
+                                            block_id,
+                                            source_mc_node_idx,
+                                            source_node_adnl_id,
+                                            self.node_debug_id,
+                                            next_allowed_send_back_time.duration_since(SystemTime::now()).unwrap_or(Duration::from_secs(0)).as_secs_f64(),
+                                        );
+                                    }
+                                }
+                            }
+                        } else {
+                            //awake block synchronization to deliver block to MC if syncrhonization phase is finished
+
+                            if !block.lock().is_synchronizing(Duration::from_millis(delivery_params.config_params.block_sync_lifetime_period_ms as u64)) {
+                                log::debug!(target: "verificator", "Awake block {:?} synchronization for workchain {} to deliver block to MC overlay", block_id, self.node_debug_id);
+                                block.lock().awake_synchronization();
+                            }
+                        }
+                    }
+                }
+
+                if !self_status_changed {
                     //block status is the same
                     return block.clone();
                 }
@@ -1073,6 +1183,12 @@ impl Workchain {
                 self.block_status_received_in_mc_counter.success();
             } else {
                 self.block_status_received_in_mc_counter.failure();
+            }
+
+            //mark delivered block as received from MC overlay
+
+            if is_delivered {
+                block.lock().mark_as_delivered_and_received_from_mc();
             }
 
             if !was_mc_processed && is_delivered {
@@ -1420,16 +1536,26 @@ impl Workchain {
         });
     }
 
-    /// Get node index by adnl id
+    /// Get WC node index by adnl id
     fn get_wc_validator_idx_by_adnl_id(&self, adnl_id: &PublicKeyHash) -> Option<usize> {
-        for (idx, src_adnl_id) in self.wc_validators_adnl_ids.iter().enumerate() {
+        self.get_validator_idx_by_adnl_id(&self.wc_validators_adnl_ids, adnl_id)
+    }
+
+    /// Get MC node index by adnl id
+    fn get_mc_validator_idx_by_adnl_id(&self, adnl_id: &PublicKeyHash) -> Option<usize> {
+        self.get_validator_idx_by_adnl_id(&self.mc_validators_adnl_ids, adnl_id)
+    }
+
+    /// Get node index by adnl id
+    fn get_validator_idx_by_adnl_id(&self, node_adnl_ids: &Vec<PublicKeyHash>, adnl_id: &PublicKeyHash) -> Option<usize> {
+        for (idx, src_adnl_id) in node_adnl_ids.iter().enumerate() {
             if src_adnl_id == adnl_id {
                 return Some(idx);
             }
         }
 
         None
-    }
+    }    
 
     /// Block status update has been received
     pub fn process_block_status(&self, adnl_id: &PublicKeyHash, block_status: BlockCandidateStatus, received_from_workchain: bool) -> Result<BlockPtr> {
@@ -1479,17 +1605,11 @@ impl Workchain {
         let approvals_signature = approvals_signature.unwrap();
         let rejections_signature = rejections_signature.unwrap();
 
-        let source_node_idx = if received_from_workchain {
-            self.get_wc_validator_idx_by_adnl_id(adnl_id)
-        } else { 
-            None
-        };
-
         //merge block status
 
         Ok(self.merge_block_status(
             &block_id,
-            source_node_idx,
+            adnl_id,
             &deliveries_signature,
             &approvals_signature,
             &rejections_signature,
@@ -1534,6 +1654,15 @@ impl Workchain {
                     block.mark_as_sent_to_mc();
                     block.get_delivery_stats().out_mc_syncs_count.fetch_add(1, Ordering::SeqCst);
                     block.get_delivery_stats().out_mc_sends_count.fetch_add(self.mc_validators.len(), Ordering::SeqCst);
+                }
+
+                {
+                    let mut block = block.lock();
+                    let now = SystemTime::now();
+    
+                    for mc_node_idx in 0..self.mc_validators.len() {
+                        block.set_mc_node_status_sent_time(mc_node_idx, now);
+                    }
                 }
 
                 self.send_block_status_to_mc(serialized_block_status);
@@ -1585,7 +1714,7 @@ impl Workchain {
             block.get_delivery_stats().out_wc_sends_count.fetch_add(nodes.len(), Ordering::SeqCst);
         
             for node_idx in nodes.iter() {
-                block.set_node_status_sent_time(*node_idx, now);
+                block.set_wc_node_status_sent_time(*node_idx, now);
             }
 
             (block.serialize(), block.get_id().clone())
@@ -1618,8 +1747,8 @@ impl Workchain {
             block.get_delivery_stats().forwarding_neighbours_sends.fetch_add(1, Ordering::SeqCst);
             block.get_delivery_stats().out_wc_sends_count.fetch_add(forwarding_neighbours.len(), Ordering::SeqCst);
         
-            for node_idx in forwarding_neighbours.iter() {
-                block.set_node_status_sent_time(*node_idx, now);
+            for wc_node_idx in forwarding_neighbours.iter() {
+                block.set_wc_node_status_sent_time(*wc_node_idx, now);
             }
 
             (block.serialize(), block.get_id().clone())
@@ -1636,7 +1765,7 @@ impl Workchain {
     fn send_block_status_to_far_neighbours(&self, block: &BlockPtr, max_neighbours_count: usize, far_neighbours_resync_period: Duration) {
         let max_far_neighbours_sync_time = std::time::SystemTime::now() - far_neighbours_resync_period;
         let wc_cutoff_weight = self.get_wc_cutoff_weight();
-        let far_neighbours = block.lock().calc_low_delivery_nodes_indexes(max_neighbours_count, wc_cutoff_weight, self.wc_local_idx as usize, max_far_neighbours_sync_time);
+        let far_neighbours = block.lock().calc_low_delivery_wc_nodes_indexes(max_neighbours_count, wc_cutoff_weight, self.wc_local_idx as usize, max_far_neighbours_sync_time);
 
         if far_neighbours.len() < 1 {
             return;
@@ -1655,7 +1784,7 @@ impl Workchain {
             block.get_delivery_stats().out_wc_sends_count.fetch_add(far_neighbours.len(), Ordering::SeqCst);
 
             for node_idx in far_neighbours.iter() {
-                block.set_node_status_sent_time(*node_idx, now);
+                block.set_wc_node_status_sent_time(*node_idx, now);
             }
 
             (block.serialize(), block.get_id().clone())
@@ -1797,7 +1926,7 @@ impl Workchain {
         if let Some(mc_overlay) = self.get_mc_overlay() {
             //convert nodes indexes to full list of validators indexes (shifted by number of MC validators)
 
-            let validators = validators.clone().iter().map(|idx| *idx + self.mc_validators.len()).collect();
+            let validators = validators.clone().iter().map(|idx| *idx + self.mc_validators.len()).collect();            
 
             mc_overlay.send_message_to_validators(data, &validators);
         }
@@ -1816,10 +1945,28 @@ impl Workchain {
         log::trace!(target: "verificator", "Workchain::send_block_status_to_mc");
 
         if let Some(mc_overlay) = self.get_mc_overlay() {
-            self.send_block_status_to_mc_counter.increment(1);
+            self.send_block_status_to_mc_counter.increment(mc_overlay.get_nodes_count() as u64);            
 
             mc_overlay.send_all(data, 0, self.mc_validators.len());
         }
+    }
+
+    /// Send block directly to specific MC node
+    fn send_block_status_to_mc_node(&self, block: &BlockPtr, mc_node_idx: usize) {
+        log::trace!(target: "verificator", "Workchain::send_block_status_to_mc_node: block={:?}, mc_node_idx={}, workchain={}", block.lock().get_id(), mc_node_idx, self.node_debug_id);
+
+        if let Some(mc_overlay) = self.get_mc_overlay() {
+            self.send_block_status_to_mc_counter.increment(1);
+
+            block.lock().get_delivery_stats().out_mc_sends_count.fetch_add(1, Ordering::SeqCst);
+
+            let serialized_block_status = block.lock().serialize();
+            let validators = vec![mc_node_idx];
+            
+            block.lock().set_mc_node_status_sent_time(mc_node_idx, SystemTime::now());
+
+            mc_overlay.send_message_to_validators(serialized_block_status, &validators);
+        }        
     }
 
     /*